requirements.txt
flask
sentence-transformers
faiss-cpu
pickle-mixin
